fabric: Merge fi_var.h into fi_prov.h
Most of the calls are intended to be used by providers only.
Move them into fi_prov.h to more clearly indicate this.
Move the fi_getsettings/fi_freesettings calls into fabric.h,
as those are intended to be used by applications (well, fi_info
anyway).
